// src/app/features/guest/models/guest-dashboard-view.model.ts
/*
 * Copyright 2021 Bundesrepublik Deutschland
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * Model containing the guest dashboard attributes in the view mode
 */
export default class GuestDashboardViewModel {
  id: string;

  firstName: string;

  lastName: string;

  arriving: string;

  leaving: string;

  email: string;

  location: string;

  status?: string;

  accreditationId?: string;
}